+++ tests/run.mjs
@@ -21,5 +21,6 @@ run([
 run([
   "--test",
   resolve(testsDir, "integration", "package-install.test.mjs"),
+  resolve(testsDir, "integration", "rule-groups.test.mjs"),
   resolve(testsDir, "integration", "svelte-parser.test.mjs"),
 ]);